feat: add GetCallLine
