+++ case_format.go
@@ -1,8 +1,6 @@
 package goutil
 
 import (
-	"reflect"
-	"runtime"
 	"strings"
 )
 
@@ -47,14 +45,4 @@ func CamelString(s string) string {
 		data = append(data, d)
 	}
 	return string(data[:])
-}
-
-// ObjectName gets the type name of the object
-func ObjectName(obj interface{}) string {
-	v := reflect.ValueOf(obj)
-	t := v.Type()
-	if t.Kind() == reflect.Func {
-		return runtime.FuncForPC(v.Pointer()).Name()
-	}
-	return t.String()
 }
