+++ trace_test.go
@@ -1,0 +1,10 @@
+package goutil
+
+import (
+	"testing"
+)
+
+func TestGetCallLine(t *testing.T) {
+	line := GetCallLine(0)
+	t.Log("====", line)
+}
